% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_classifier.R
\name{beta_classifier}
\alias{beta_classifier}
\title{Classification using ONLY optimal scaling}
\usage{
beta_classifier(yt, group, L, yt_new)
}
\value{
classes
}
\description{
Classification using ONLY optimal scaling
}
\examples{
To be added later
}
