% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_get.R
\name{env_get}
\alias{env_get}
\title{Given a m times p-1 categorical time series, compute:
(1) the Fourier frequencies
(2) the spectral envelope
(3) the optimal scaling}
\usage{
env_get(yt, L)
}
\arguments{
\item{yt}{DESCRIPTION}

\item{L}{DESCRIPTION}
}
\value{
DESCRIPTION
}
\description{
Given a m times p-1 categorical time series, compute:
(1) the Fourier frequencies
(2) the spectral envelope
(3) the optimal scaling
}
\examples{
To be added later
}
