% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_classifier.R
\name{gamma_classifier}
\alias{gamma_classifier}
\title{Classification using ONLY spectral envelope}
\usage{
gamma_classifier(yt, group, L, yt_new)
}
\value{
classes
}
\description{
Classification using ONLY spectral envelope
}
\examples{
To be added later
}
