% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_classifier.R
\name{beta_classifier}
\alias{beta_classifier}
\title{Classification Using only Optimal Scaling}
\usage{
beta_classifier(yt, group, L, yt_new)
}
\arguments{
\item{yt}{A 3D array representing the training time-series. Each slice corresponds to one time-series.}

\item{group}{A vector of integers indicating the class/group for each time-series in \code{yt}. The
length of \code{group} must match the number of slices in \code{yt}.}

\item{L}{A integer or vector of integers giving the widths of modified Daniell smoothers to be
used to smooth the periodogram. The value of the elements in \code{L} range from 1 to cube root of the number
of rows of \code{yt}.}

\item{yt_new}{A matrix representing the test time series. Dimensions are the same as \code{yt}.}
}
\value{
A numeric vector where each element corresponds to the predicted class label for a test time-series in
\code{yt_new}.
}
\description{
This function classifies test time-series based solely on the optimal scaling matrices derived from the
training time-series groups.
}
\examples{
set.seed(12092024)

# Simulate training time-series for two classes
yt <- array(rnorm(1500), dim = c(50, 3, 10))
group <- c(rep(1, 5), rep(2, 5))

# Simulate test time-series
yt_new <- matrix(rnorm(150), nrow = 50)

# Classify the test time-series
classes <- beta_classifier(yt, group, L = 3, yt_new)
print(classes)

}
