% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_convert.R
\name{cat_convert}
\alias{cat_convert}
\title{Convert a categorical time-series into multivariate indicator process}
\usage{
cat_convert(xt)
}
\arguments{
\item{xt}{DESCRIPTION}
}
\value{
DESCRIPTION
}
\description{
Convert a categorical time-series into multivariate indicator process
}
\examples{
To be added later
}
