% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_classifier.R
\name{env_classifier}
\alias{env_classifier}
\title{This is the main function: It assigns classes to a time series}
\usage{
env_classifier(yt, group, L, yt_new, kappa)
}
\arguments{
\item{yt}{DESCRIPTION}

\item{group}{DESCRIPTION}

\item{L}{DESCRIPTION}

\item{yt_new}{DESCRIPTION}

\item{kappa}{DESCRIPTION}
}
\value{
classes
}
\description{
This is the main function: It assigns classes to a time series
}
\examples{
To be added later
}
