% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_classifier_crossv.R
\name{env_classifier_crossv}
\alias{env_classifier_crossv}
\title{Cross-validation to do grid search of kappa}
\usage{
env_classifier_crossv(yt, group, L, kappa)
}
\value{
classes
}
\description{
Cross-validation to do grid search of kappa
}
\examples{
To be added later
}
