% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_get.R
\name{env_get}
\alias{env_get}
\title{Get frequencies, spectral envelope, and optimal scaling.}
\usage{
env_get(yt, L)
}
\arguments{
\item{yt}{A matrix containing univariate or multivariate time-series. Rows represent time points.}

\item{L}{A integer or vector of integers giving the widths of modified Daniell smoothers to be
used to smooth the periodogram. The value of the elements in \code{L} range from 1 to cube root of the number
of rows of \code{yt}.}
}
\value{
A list containing the following components:
\describe{
\item{\code{freq}}{A vector of frequencies.}
\item{\code{envelope}}{A vector of spectral envelope values corresponding to the frequencies.}
\item{\code{scale}}{A matrix of optimal scaling vectors corresponding to the frequencies.}
}
}
\description{
Given a m times p-1 categorical time series, compute:
(1) the Fourier frequencies
(2) the spectral envelope
(3) the optimal scaling
}
\examples{
set.seed(12092024)
data <- matrix(rnorm(500), ncol = 5)
result <- env_get(data, L = 3)

# Accessing results
freq <- result$freq # Frequencies
envelope <- result$envelope # Spectral Envelope
scale <- result$scale # Optimal Scaling

}
