% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_classifier.R
\name{gamma_classifier}
\alias{gamma_classifier}
\title{Classification Using only Spectral Envelope}
\usage{
gamma_classifier(yt, group, L, yt_new, plot = FALSE)
}
\arguments{
\item{yt}{A 3D array representing the training time-series. Each slice corresponds to one time-series.}

\item{group}{A vector of integers indicating the class/group for each time-series in \code{yt}. The
length of \code{group} must match the number of slices in \code{yt}.}

\item{L}{A integer or vector of integers giving the widths of modified Daniell smoothers to be
used to smooth the periodogram. The value of the elements in \code{L} range from 2 to less than half of the
number of rows of \code{yt}. It is feasible to have \code{L} less than the cube root of the number of
rows in \code{yt}.}

\item{yt_new}{A matrix or array representing the test time series. Dimensions for number of rows and columns
are the same as \code{yt}.}

\item{plot}{Logical; If \code{TRUE}, generates plots for group spectral envelopes. Default is \code{FALSE}.}
}
\value{
A numeric vector where each element corresponds to the predicted class label for a test time-series in
\code{yt_new}.
}
\description{
This function classifies test time-series based solely on the spectral envelope derived from the
training time-series groups.
}
\examples{
# Example: 1
set.seed(12092024)

# Simulate training time-series for two classes
yt <- array(rnorm(1500), dim = c(50, 3, 10))
group <- c(rep(1, 5), rep(2, 5))

# Simulate test time-series
yt_new <- array(rnorm(600), dim = c(50, 3, 4))

# Classify the test time-series
classes <- gamma_classifier(yt, group, L = 3, yt_new, TRUE)
print(classes)

# Example: 2
set.seed(12092024)

# Simulate training time-series for two classes
yt <- array(rnorm(1500), dim = c(50, 3, 10))
group <- c(rep(1, 5), rep(2, 5))

# Simulate test time-series
yt_new <- array(rnorm(150), dim = c(50, 3, 1))

# Classify the test time-series
classes <- gamma_classifier(yt, group, L = 3, yt_new, TRUE)
print(classes)

# Example: 3
set.seed(12092024)

# Simulate training time-series for two classes
yt <- array(rnorm(1500), dim = c(50, 3, 10))
group <- c(rep(1, 5), rep(2, 5))

# Simulate test time-series
yt_new <- matrix(rnorm(150), nrow = 50)

# Classify the test time-series
classes <- gamma_classifier(yt, group, L = 3, yt_new, TRUE)
print(classes)

}
