% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_env.R
\name{group_env}
\alias{group_env}
\title{Given a several time series within a group
Compute
(1) the group level spectral envelope
(2) the group level optimal scaling}
\usage{
group_env(yt_group, L)
}
\description{
Given a several time series within a group
Compute
(1) the group level spectral envelope
(2) the group level optimal scaling
}
\examples{
To be added later
}
